生产环境配置，用户层级类型改为long

+++ crius-service/src/main/java/com/magic/crius/dao/crius/db/UserInfoMapper.java
@@ -30,6 +30,6 @@ public interface UserInfoMapper {
      * @param level
      * @return
      */
-    int updateLevel(@Param("userId") Long userId,@Param("userLevel") Integer level);
+    int updateLevel(@Param("userId") Long userId,@Param("userLevel") Long level);
 
 }
